requirements.txt
requests
pillow
idotmatrix
bleak
numpy
python-dotenv
